Move GA4 sign_up logic to getTokenFromSSO
